Add Python code generator, native scalar hash and avx512 benchmark implementations
Removal of simd_main.cpp allowed for splitting the different functionalities into separate files. This change is intended to improve the extensibility of the project by generating specific code sequences via the new Python script and introducing scalar and avx512 benchmarking capabilities with the new hash files. Modifications to CMakeLists.txt are meant to allow for more specific compilation options to cater to different use cases.